move wrap_run to tabular
